requirements.txt
ultralytics
gradio
opencv-python
Pillow
pyyaml
pandas
